Move log formatting etc. to root logger
Underlying libraries will also log things, while this is a client
program.  It doesn't make sense to exclusively to set up global
log levels and formatting just for the zinoargus logger; the formatting
should apply to all log output.

Also, it does not make sense to set a log level both for the stdout
handler *and* the logger.  It's fine to set it for only one of them,
and I prefer logger-centric levels.