Clean up metadata collection naming
This removes useless comments by renaming the `collect_metadata()`
function to `collect_circuit_metadata()` (also renaming the global
variable).

It's still not clear what the circuit metadata was supposed to be
used for, since development of this code was halted for three years.
Whatever it means, it should probably be outsourced to something
more pluggable, as it seems very Uninett-specific.