Revamp continuous update logic
There was a bunch of strange repetition in multiple branches of the Zino
notification loop.  This simplifies the logic somewhat and extracts
a function for handling state change notifications.

Basically, any time we receive an update about a Zino case, it doesn't
matter what kind of update it is.  We need to first always get updated
data on the case it's referring to, and if that case is "interesting"
enough to push to Argus, we should get an existing incident or make
a new one.

Updates can also come for Zino cases that have been recently closed,
so the corresponding Argus incident, if any, may also have been closed,
so we may still need to fetch a closed Argus incident if we no
longer have it in internal data structures.

Acting on the different kinds of notifications that come in from Zino
is the next step after this common functionality, and the various
update types should be factored out to separate sub-functions.