Sync Zino case history to Argus incident events
This implements an idempotent Zino case history synchronization
function and calls it on history update notifications from Zino.

Zino notifications only tell us the case history has been updated, but
not which entries have been added.  Since there is no identifying key
other than timestamps that we can use to correlate Zino history entries
with Argus incident event entries, the best approach seems to be to make
a full correlation of the full log on both sides whenever an update is
received and post only those entries which seem to not be in Argus
already.