Set proper timestamp/log when closing incidents
As per the TODO item in the comments, we need to use a proper timestamp
when resolving an Argus incident.  The current time isn't good enough,
as we may be processing the closure some time after the Zino case was
closed.

There is no official "closing time" in Zino cases, so the best we can
do is fetch the timestamp of the last history entry of a closed Zino
case, and use that as the timestamp sent to Zino.

Additionally, the message/header from the last history entry is a much
better description of the incident closure than the hardcoded
"Zino case closed by user", so this change also allows for the
closure event description to be taken from the zino case.